Dev/products/Idea_08_Meeting_Commitment_Extractor/src/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
pydantic-settings>=2.1.0
deepgram-sdk>=3.0.0
anthropic>=0.18.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
asana>=3.2.0
slack-sdk>=3.27.0
